sandbox/seccomp: Use kill instead of trap
In the past we used `Trap` as default action for being able to print the
failed syscall to STDERR in a signal handler. However, in most cases the
filtered syscall was already used before the signal handler was connected in
ctors. Hence this approach wasn't very successful.

Using `KillProcess` at least allows this event to be logged by auditd.